+++ releaf-i18n_database/lib/releaf/i18n_database/backend.rb
@@ -6,7 +6,7 @@ module Releaf
 
       include ::I18n::Backend::Base, ::I18n::Backend::Flatten
       UPDATED_AT_KEY = 'releaf.i18n_database.translations.updated_at'
-      DEFAULT_CONFIG = {auto_creation: true, auto_creation_exception_patterns: [/^attributes\./]}
+      DEFAULT_CONFIG = {translation_auto_creation: true, translation_auto_creation_exclusion_patterns: [/^attributes\./]}
       attr_accessor :translations_cache
 
       def self.initialize_component