+++ releaf-i18n_database/lib/releaf/i18n_database/configuration.rb
@@ -1,7 +1,7 @@
 module Releaf::I18nDatabase
   class Configuration
     include Virtus.model(strict: true)
-    attribute :auto_creation, Boolean
-    attribute :auto_creation_exception_patterns, Array
+    attribute :translation_auto_creation, Boolean
+    attribute :translation_auto_creation_exclusion_patterns, Array
   end
 end