+++ releaf-i18n_database/spec/features/translations_spec.rb
@@ -195,7 +195,7 @@ feature "Translations" do
   describe "Lookup" do
     background do
       Releaf::I18nDatabase::Backend.reset_cache
-      allow( Releaf.application.config.i18n_database ).to receive(:auto_creation).and_return(true)
+      allow( Releaf.application.config.i18n_database ).to receive(:translation_auto_creation).and_return(true)
     end
 
     context "when translation exists within higher level key (instead of being scope)" do