+++ releaf-i18n_database/spec/lib/releaf/i18n_database/configuration_spec.rb
@@ -1,11 +1,11 @@
 require "rails_helper"
 
 describe Releaf::I18nDatabase::Configuration do
-  subject{ described_class.new(auto_creation: true, auto_creation_exception_patterns: [1, 2]) }
+  subject{ described_class.new(translation_auto_creation: true, translation_auto_creation_exclusion_patterns: [1, 2]) }
 
   it do
-    is_expected.to have_attributes(auto_creation: true)
-    is_expected.to have_attributes(auto_creation_exception_patterns: [1, 2])
+    is_expected.to have_attributes(translation_auto_creation: true)
+    is_expected.to have_attributes(translation_auto_creation_exclusion_patterns: [1, 2])
   end
 end