+++ spec/rails_helper.rb
@@ -95,7 +95,7 @@ RSpec.configure do |config|
 
   config.before(:each) do |example|
     Rails.cache.clear
-    allow( Releaf.application.config.i18n_database ).to receive(:auto_creation).and_return(false)
+    allow( Releaf.application.config.i18n_database ).to receive(:translation_auto_creation).and_return(false)
 
     if example.metadata[:db_strategy]
       DatabaseCleaner.strategy = example.metadata[:db_strategy]